+++ lib/templatechecker/class.WidgetParameterChecker.php
@@ -462,12 +462,12 @@ class WidgetParameterChecker {
 			return;
 
 		// basic.window and device.window and maybe other dynamic icons in future have a special color mode (variable / constant)
-		// constant mode needs to start with '!#' and be 5 (!#+3) or 8 (!#+6) characters long
-		if ($this->checkParameterValidValues('!') && substr($value, 0, 2) == '!#' && (strlen($value) == 5 || strlen($value) == 8) && ctype_xdigit(substr($value, 2)))
+		// constant mode needs to start with '!#' and be 5 (!#+3), 6, 8 or 10 (!#+8) characters long including RBGA colors
+		if ($this->checkParameterValidValues('!') && substr($value, 0, 2) == '!#' && (strlen($value) == 5 || strlen($value) == 6 ||strlen($value) == 8 ||strlen($value) == 10) && ctype_xdigit(substr($value, 2)))
 			return;
 		
-		// anything else needs to start with '#' and be 4 (#+3) or 7 (#+6) characters long
-		if (substr($value, 0, 1) == '#' && (strlen($value) == 4 || strlen($value) == 7) && ctype_xdigit(substr($value, 1)))
+		// anything else needs to start with '#' and be 4 (#+3), 5, 7 or 9 (#+8) characters long including RBGA colors
+		if (substr($value, 0, 1) == '#' && (strlen($value) == 4 || strlen($value) == 5 ||strlen($value) == 7 ||strlen($value) == 9) && ctype_xdigit(substr($value, 1)))
 			return;
 
 		// TODO: rgb() / rgba() / hsl()